Close file handle in preloadFont
Commit 785f90a7f546f0e94a40ad0e9cc3e5e64322805d introduced a regression
by not using a context manager when opening the file, which eventually
lead to a resource leak because the file handle 'f' never gets closed.

This is fixed by passing the font path, instead of the open file handle,
and by opening the handle using a 'with' context manager.

The issue can be reproduced by running pytest with the '-W error' flag
for the following minimal example:

    def test_file_gets_closed():
        Figlet().renderText("")

This will succeed now, but fails for v1.0.1.